move scp330 event to right namespace
