+++ Exiled.Events/Handlers/Scp330.cs
@@ -7,7 +7,6 @@
 
 namespace Exiled.Events.Handlers
 {
-    using Exiled.Events.EventArgs.Player;
     using Exiled.Events.EventArgs.Scp330;
     using Extensions;
